Task #1418: fix various MAC protocols to restore compat with 64 bit Linux (size_t -> uint32). Most of this stuff elsewhere was already patched earlier by Ruud, and some supplemental fixes by me, but apparently there is more and this stuff is not in jenkins(?). The LOFAR tree now appears to build (I do that without PVSS).

+++ trunk/MAC/APL/PAC/CAL_Protocol/src/SpectralWindow.cc
@@ -148,7 +148,7 @@ size_t SpectralWindow::getSize() const
 
 size_t SpectralWindow::pack(char* buffer) const
 {
-  size_t offset = 0;
+  uint32 offset = 0;
 
   MSH_pack(buffer, offset, m_name);
   memcpy(buffer + offset, &m_sampling_freq, sizeof(m_sampling_freq));
@@ -165,7 +165,7 @@ size_t SpectralWindow::pack(char* buffer) const
 
 size_t SpectralWindow::unpack(const char* buffer)
 {
-  size_t offset = 0;
+  uint32 offset = 0;
 
   MSH_unpack(buffer, offset, m_name);
   memcpy(&m_sampling_freq, buffer + offset, sizeof(m_sampling_freq));
